+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.slice 0.5.1.9005
 
+### Enhancements
+* The `ui_add` and `srv_add` methods from the `FilteredData` class are being called inside the `ui_active` and `srv_active` methods, respectively. This allows us to reduce the space of the filter panel by not displaying the "add filters" UI in a separate panel.
+
 ### Bug fixes
 * Fix error while creating the filter choices when the data has a factor with a level containing an empty string ("").